add hard reset

+++ include/Packet.h
@@ -44,7 +44,7 @@ struct packet_t{
   unsigned int           DATA_FOOTER[numChipsOnBoard];
   /* raw data saved here */
   unsigned short         AC_RAW_DATA[numChipsOnBoard][psec_buffersize];
-  float                  Data[AC_CHANNELS+1][psecSampleCells]; /* AC_CHANNELS waveforms + 1 metadata */
+  unsigned short         Data[AC_CHANNELS+1][psecSampleCells]; /* AC_CHANNELS waveforms + 1 metadata */
   unsigned short         AC_INFO[numChipsOnBoard][infoBuffersize];
 };
